fix(ollama): inconsistent package name
