+++ prompt/prompt-executor/prompt-executor-clients/prompt-executor-ollama-client/src/jsMain/kotlin/ai/koog/prompt/executor/clients/ollama/OllamaClient.js.kt
@@ -1,4 +1,4 @@
-package ai.koog.prompt.executor.ollama.client
+package ai.koog.prompt.executor.clients.ollama
 
 import io.ktor.client.engine.HttpClientEngineFactory
 import io.ktor.client.engine.js.Js
