+++ prompt/prompt-executor/prompt-executor-clients/prompt-executor-ollama-client/src/jvmMain/kotlin/ai/koog/prompt/executor/clients/ollama/OllamaClient.jvm.kt
@@ -1,4 +1,4 @@
-package ai.koog.prompt.executor.ollama.client
+package ai.koog.prompt.executor.clients.ollama
 
 import io.ktor.client.engine.HttpClientEngineFactory
 import io.ktor.client.engine.cio.CIO
